requirements.txt
python-telegram-bot[webhooks]>=20.0
python-dotenv
pymongo[srv]
google-generativeai
flask
requests
